Refactoring to extract more discrete classes
