+++ src/main/java/org/sonatype/goodies/packageurl/jaxb/PackageUrlXmlAdapter.java
@@ -16,7 +16,7 @@ import javax.annotation.Nullable;
 import javax.xml.bind.annotation.adapters.XmlAdapter;
 
 import org.sonatype.goodies.packageurl.PackageUrl;
-import org.sonatype.goodies.packageurl.PackageUrl.RenderFlavor;
+import org.sonatype.goodies.packageurl.RenderFlavor;
 
 import static com.google.common.base.Preconditions.checkNotNull;
